+++ src/ONTraC/analysis/cell_type.py
@@ -14,6 +14,7 @@ import seaborn as sns
 
 from ..log import warning
 from .data import AnaData
+from .niche_cluster import cal_nc_order, cal_nc_order_index, cal_nc_scores
 from .utils import saptial_figsize
 
 
@@ -161,12 +162,8 @@ def plot_violin_cell_type_along_NT_score_from_anadata(ana_data: AnaData,
     :return: None or Tuple[plt.Figure, plt.Axes].
     """
 
-    try:
-        if 'Cell_NTScore' not in ana_data.NT_score.columns:
-            warning("No NT score data found.")
-            return None
-    except FileNotFoundError as e:
-        warning(str(e))
+    if ana_data.NT_score is None:
+        warning("No NT score data found.")
         return None
 
     data_df = ana_data.meta_data_df.join(ana_data.NT_score['Cell_NTScore'])
@@ -209,12 +206,8 @@ def plot_kde_cell_type_along_NT_score_from_anadata(ana_data: AnaData) -> Optiona
     :return: None or Tuple[plt.Figure, plt.Axes].
     """
 
-    try:
-        if 'Cell_NTScore' not in ana_data.NT_score.columns:
-            warning("No NT score data found.")
-            return None
-    except FileNotFoundError as e:
-        warning(str(e))
+    if ana_data.NT_score is None:
+        warning("No NT score data found.")
         return None
 
     data_df = ana_data.meta_data_df.join(ana_data.NT_score['Cell_NTScore'])
@@ -254,12 +247,8 @@ def plot_hist_cell_type_along_NT_score_from_anadata(ana_data: AnaData) -> Option
     :return: None or Tuple[plt.Figure, plt.Axes].
     """
 
-    try:
-        if 'Cell_NTScore' not in ana_data.NT_score.columns:
-            warning("No NT score data found.")
-            return None
-    except FileNotFoundError as e:
-        warning(str(e))
+    if ana_data.NT_score is None:
+        warning("No NT score data found.")
         return None
 
     data_df = ana_data.meta_data_df.join(ana_data.NT_score['Cell_NTScore'])
@@ -279,12 +268,8 @@ def plot_cell_type_along_NT_score(ana_data: AnaData) -> None:
     :return: None.
     """
 
-    try:
-        if 'Cell_NTScore' not in ana_data.NT_score.columns:
-            warning("No NT score data found.")
-            return None
-    except FileNotFoundError as e:
-        warning(str(e))
+    if ana_data.NT_score is None:
+        warning("No NT score data found.")
         return None
 
     data_df = ana_data.meta_data_df.join(ana_data.NT_score['Cell_NTScore'])
@@ -343,15 +328,11 @@ def plot_cell_type_loading_in_niche_clusters_from_anadata(ana_data: AnaData) -> 
     :return: None or Tuple[plt.Figure, plt.Axes].
     """
 
-    try:
-        if ana_data.cell_level_niche_cluster_assign is None:
-            warning("No niche cluster assign data found.")
-            return None
-        if ana_data.cell_type_codes is None:
-            warning("No cell type data found.")
-            return None
-    except FileNotFoundError as e:
-        warning(str(e))
+    if ana_data.cell_level_niche_cluster_assign is None:
+        warning("No niche cluster assign data found.")
+        return None
+    if ana_data.cell_type_codes is None:
+        warning("No cell type data found.")
         return None
 
     # calculate cell type distribution in each niche cluster
@@ -367,7 +348,9 @@ def plot_cell_type_loading_in_niche_clusters_from_anadata(ana_data: AnaData) -> 
     if ana_data.options.output is not None:
         cell_type_dis_df.to_csv(f'{ana_data.options.output}/cell_type_dis_in_niche_clusters.csv', index=False)
     # nc_order
-    nc_scores = 1 - ana_data.niche_cluster_score if ana_data.options.reverse else ana_data.niche_cluster_score
+    nc_scores = cal_nc_scores(cell_level_niche_cluster_assign=ana_data.cell_level_niche_cluster_assign,
+                              reverse=ana_data.options.reverse,
+                              niche_cluster_score=ana_data.niche_cluster_score)
     nc_order = [f'NicheCluster_{x}' for x in nc_scores.argsort()]
     cell_type_dis_df = cell_type_dis_df.loc[nc_order]
 
@@ -375,44 +358,41 @@ def plot_cell_type_loading_in_niche_clusters_from_anadata(ana_data: AnaData) -> 
                                                     output_file_path=ana_data.options.output)
 
 
-def plot_cell_type_dis_in_niche_clusters(
+def plot_cell_type_com_in_niche_clusters(
         cell_type_dis_df: pd.DataFrame,
         output_file_path: Optional[Union[str, Path]] = None) -> Optional[Tuple[plt.Figure, plt.Axes]]:
     """
-    Plot cell type distribution in each niche cluster.
+    Plot cell type composition in each niche cluster.
     :param ana_data: AnaData, the data for analysis.
     :param cell_type_dis_df: pd.DataFrame, the cell type distribution in each niche cluster.
     :return: None or Tuple[plt.Figure, plt.Axes]
     """
+
     fig, ax = plt.subplots(figsize=(2 + cell_type_dis_df.shape[1] / 3, 1 + cell_type_dis_df.shape[0] / 5))
     sns.heatmap(cell_type_dis_df.apply(lambda x: x / x.sum(), axis=1), ax=ax)
     ax.set_xlabel('Cell Type')
     ax.set_ylabel('Niche Cluster')
     fig.tight_layout()
     if output_file_path is not None:
-        fig.savefig(f'{output_file_path}/cell_type_dis_in_niche_clusters.pdf', transparent=True)
+        fig.savefig(f'{output_file_path}/cell_type_composition_in_niche_clusters.pdf', transparent=True)
         plt.close(fig)
         return None
     else:
         return fig, ax
 
 
-def plot_cell_type_dis_in_niche_clusters_from_anadata(ana_data: AnaData) -> Optional[Tuple[plt.Figure, plt.Axes]]:
+def plot_cell_type_com_in_niche_clusters_from_anadata(ana_data: AnaData) -> Optional[Tuple[plt.Figure, plt.Axes]]:
     """
-    Plot cell type distribution in each niche cluster.
+    Plot cell type composition in each niche cluster.
     :param ana_data: AnaData, the data for analysis.
     :return: None or Tuple[plt.Figure, plt.Axes].
     """
 
-    try:
-        if ana_data.cell_level_niche_cluster_assign is None:
-            warning("No niche cluster assign data found.")
-            return None
-        if ana_data.cell_type_codes is None:
-            warning("No cell type data found.")
-            return None
-    except FileNotFoundError as e:
-        warning(str(e))
+    if ana_data.cell_level_niche_cluster_assign is None:
+        warning("No niche cluster assign data found.")
+        return None
+    if ana_data.cell_type_codes is None:
+        warning("No cell type data found.")
         return None
 
     # calculate cell type distribution in each niche cluster
@@ -428,15 +408,18 @@ def plot_cell_type_dis_in_niche_clusters_from_anadata(ana_data: AnaData) -> Opti
     if ana_data.options.output is not None:
         cell_type_dis_df.to_csv(f'{ana_data.options.output}/cell_type_dis_in_niche_clusters.csv', index=False)
     # nc_order
-    nc_scores = 1 - ana_data.niche_cluster_score if ana_data.options.reverse else ana_data.niche_cluster_score
-    nc_order = [f'NicheCluster_{x}' for x in nc_scores.argsort()]
+    nc_scores = cal_nc_scores(cell_level_niche_cluster_assign=ana_data.cell_level_niche_cluster_assign,
+                              reverse=ana_data.options.reverse,
+                              niche_cluster_score=ana_data.niche_cluster_score)
+    nc_order = [f'NicheCluster_{x.split()[-1]}' for x in cal_nc_order(cal_nc_order_index(nc_scores))]
+    # cell type distribution dataframe
     cell_type_dis_df = cell_type_dis_df.loc[nc_order]
 
-    return plot_cell_type_dis_in_niche_clusters(cell_type_dis_df=cell_type_dis_df,
+    return plot_cell_type_com_in_niche_clusters(cell_type_dis_df=cell_type_dis_df,
                                                 output_file_path=ana_data.options.output)
 
 
-def plot_cell_type_across_niche_cluster(
+def plot_cell_type_dis_across_niche_cluster(
         cell_type_dis_df: pd.DataFrame,
         output_file_path: Optional[Union[str, Path]] = None) -> Optional[Tuple[plt.Figure, plt.Axes]]:
     """
@@ -445,6 +428,7 @@ def plot_cell_type_across_niche_cluster(
     :param cell_type_dis_df: pd.DataFrame, the cell type distribution in each niche cluster.
     :return: None or Tuple[plt.Figure, plt.Axes]
     """
+
     fig, ax = plt.subplots(figsize=(2 + cell_type_dis_df.shape[1] / 3, 1 + cell_type_dis_df.shape[0] / 5))
     sns.heatmap(cell_type_dis_df.apply(lambda x: x / x.sum(), axis=0), ax=ax)
     ax.set_xlabel('Cell Type')
@@ -458,22 +442,18 @@ def plot_cell_type_across_niche_cluster(
         return fig, ax
 
 
-def plot_cell_type_across_niche_cluster_from_anadata(ana_data: AnaData) -> Optional[Tuple[plt.Figure, plt.Axes]]:
+def plot_cell_type_dis_across_niche_cluster_from_anadata(ana_data: AnaData) -> Optional[Tuple[plt.Figure, plt.Axes]]:
     """
     Plot cell type distribution across niche cluster.
     :param ana_data: AnaData, the data for analysis.
     :return: None or Tuple[plt.Figure, plt.Axes].
     """
 
-    try:
-        if ana_data.cell_level_niche_cluster_assign is None:
-            warning("No niche cluster assign data found.")
-            return None
-        if ana_data.cell_type_codes is None:
-            warning("No cell type data found.")
-            return None
-    except FileNotFoundError as e:
-        warning(str(e))
+    if ana_data.cell_level_niche_cluster_assign is None:
+        warning("No niche cluster assign data found.")
+        return None
+    if ana_data.cell_type_codes is None:
+        warning("No cell type data found.")
         return None
 
     # calculate cell type distribution in each niche cluster
@@ -489,11 +469,14 @@ def plot_cell_type_across_niche_cluster_from_anadata(ana_data: AnaData) -> Optio
     if ana_data.options.output is not None:
         cell_type_dis_df.to_csv(f'{ana_data.options.output}/cell_type_dis_in_niche_clusters.csv', index=False)
     # nc_order
-    nc_scores = 1 - ana_data.niche_cluster_score if ana_data.options.reverse else ana_data.niche_cluster_score
-    nc_order = [f'NicheCluster_{x}' for x in nc_scores.argsort()]
+    nc_scores = cal_nc_scores(cell_level_niche_cluster_assign=ana_data.cell_level_niche_cluster_assign,
+                              reverse=ana_data.options.reverse,
+                              niche_cluster_score=ana_data.niche_cluster_score)
+    nc_order = [f'NicheCluster_{x.split()[-1]}' for x in cal_nc_order(cal_nc_order_index(nc_scores))]
+    # cell type distribution dataframe
     cell_type_dis_df = cell_type_dis_df.loc[nc_order]
 
-    return plot_cell_type_across_niche_cluster(cell_type_dis_df=cell_type_dis_df,
+    return plot_cell_type_dis_across_niche_cluster(cell_type_dis_df=cell_type_dis_df,
                                                output_file_path=ana_data.options.output)
 
 
@@ -504,18 +487,14 @@ def plot_cell_type_with_niche_cluster(ana_data: AnaData) -> None:
     :return: None.
     """
 
-    try:
-        if ana_data.cell_level_niche_cluster_assign is None:
-            warning("No niche cluster assign data found.")
-            return None
-        if ana_data.cell_type_codes is None:
-            warning("No cell type data found.")
-            return None
-        if ana_data.niche_cluster_score is None:
-            warning("No niche cluster scores data found.")
-            return None
-    except FileNotFoundError as e:
-        warning(str(e))
+    if ana_data.cell_level_niche_cluster_assign is None:
+        warning("No niche cluster assign data found.")
+        return None
+    if ana_data.cell_type_codes is None:
+        warning("No cell type data found.")
+        return None
+    if ana_data.niche_cluster_score is None:
+        warning("No niche cluster scores data found.")
         return None
 
     # calculate cell type distribution in each niche cluster
@@ -531,15 +510,17 @@ def plot_cell_type_with_niche_cluster(ana_data: AnaData) -> None:
     if ana_data.options.output is not None:
         cell_type_dis_df.to_csv(f'{ana_data.options.output}/cell_type_dis_in_niche_clusters.csv', index=False)
     # nc_order
-    nc_scores = 1 - ana_data.niche_cluster_score if ana_data.options.reverse else ana_data.niche_cluster_score
-    nc_order = [f'NicheCluster_{x}' for x in nc_scores.argsort()]
+    nc_scores = cal_nc_scores(cell_level_niche_cluster_assign=ana_data.cell_level_niche_cluster_assign,
+                              reverse=ana_data.options.reverse,
+                              niche_cluster_score=ana_data.niche_cluster_score)
+    nc_order = [f'NicheCluster_{x.split()[-1]}' for x in cal_nc_order(cal_nc_order_index(nc_scores))]
+    # cell type distribution dataframe
     cell_type_dis_df = cell_type_dis_df.loc[nc_order]
 
-    # plot_cell_type_loading_in_niche_clusters(ana_data=ana_data, cell_type_dis_df=cell_type_dis_df, nc_order=nc_order)
     plot_cell_type_loading_in_niche_clusters(cell_type_dis_df=cell_type_dis_df,
                                              output_file_path=ana_data.options.output)
-    plot_cell_type_dis_in_niche_clusters(cell_type_dis_df=cell_type_dis_df, output_file_path=ana_data.options.output)
-    plot_cell_type_across_niche_cluster(cell_type_dis_df=cell_type_dis_df, output_file_path=ana_data.options.output)
+    plot_cell_type_com_in_niche_clusters(cell_type_dis_df=cell_type_dis_df, output_file_path=ana_data.options.output)
+    plot_cell_type_dis_across_niche_cluster(cell_type_dis_df=cell_type_dis_df, output_file_path=ana_data.options.output)
 
 
 def cell_type_visualization(ana_data: AnaData) -> None: