+++ src/ONTraC/analysis/data.py
@@ -1,25 +1,25 @@
 import os
 from optparse import Values
-from typing import Dict, Union
+from typing import Dict, Optional
 
 import numpy as np
 import pandas as pd
 from pandas import DataFrame
 
-from ..log import info
+from ..log import warning
 from ..utils import get_meta_data_file, get_rel_params, read_yaml_file
 
 
 # ----------------------------
 # Misc Functions
 # ----------------------------
-def load_loss_record_data(options) -> Union[Dict, None]:
+def load_loss_record_data(options) -> Optional[Dict]:
     """
     Parse the log file and save the training loss to a csv file.
     Args:
         options: Values, the options from optparse
     Returns:
-        Dict: {'loss_df': pd.DataFrame, 'loss_dict': Dict}
+        Dict: {'loss_df': pd.DataFrame, 'loss_dict': Dict} or None
         loss_df is the training loss dataframe
         loss_dict is the final loss dictionary
 
@@ -30,7 +30,7 @@ def load_loss_record_data(options) -> Union[Dict, None]:
     if options.log is None:
         return None
     with open(options.log, 'r') as fhd:
-        for line in fhd:
+        for line in fhd: # type: ignore
             # train loss record
             if 'INFO' in line and 'epoch' in line and 'loss' in line:
                 loss_ = []
@@ -70,111 +70,114 @@ def load_loss_record_data(options) -> Union[Dict, None]:
     return {'loss_df': epoch_loss_df, 'loss_dict': final_loss_dict}
 
 
-def load_niche_cluster_connectivity(options: Values) -> np.ndarray:
+def load_niche_cluster_connectivity(options: Values) -> Optional[np.ndarray]:
     """
     Load the niche cluster connectivity from the output of GNN.
     Args:
         options: Values, the options from optparse
     Returns:
-        np.ndarray: the niche cluster connectivity
+        Optional[np.ndarray]: the niche cluster connectivity
     """
     niche_cluster_conn_file = f'{options.GNN_dir}/consolidate_out_adj.csv.gz'
     if not os.path.isfile(niche_cluster_conn_file):
         niche_cluster_conn_file = f'{options.GNN_dir}/consolidate_out_adj.csv'
     if not os.path.isfile(niche_cluster_conn_file):  # skip if file not exist
-        raise FileNotFoundError(f"Cannot find niche cluster connectivity file: {niche_cluster_conn_file}.")
+        warning(f"Cannot find niche cluster connectivity file: {niche_cluster_conn_file}.")
+        return None
 
     return np.loadtxt(f'{niche_cluster_conn_file}', delimiter=',')
 
 
-def load_niche_cluster_score(options: Values) -> np.ndarray:
+def load_niche_cluster_score(options: Values) -> Optional[np.ndarray]:
     """
     Load the niche cluster score from the output of NT score.
     Args:
         options: Values, the options from optparse
     Returns:
-        np.ndarray: the niche cluster score
+        Optional[np.ndarray]: the niche cluster score
     """
     niche_cluster_score_file = f'{options.NT_dir}/niche_cluster_score.csv.gz'
     if not os.path.isfile(niche_cluster_score_file):
         niche_cluster_score_file = f'{options.NT_dir}/niche_cluster_score.csv'
     if not os.path.isfile(niche_cluster_score_file):  # skip if file not exist
-        raise FileNotFoundError(f"Cannot find niche cluster score file: {niche_cluster_score_file}.")
+        warning(f"Cannot find niche cluster score file: {niche_cluster_score_file}.")
+        return None
 
     return np.loadtxt(f'{niche_cluster_score_file}', delimiter=',')
 
 
-def load_niche_level_niche_cluster_assign(options: Values) -> pd.DataFrame:
+def load_niche_level_niche_cluster_assign(options: Values) -> Optional[pd.DataFrame]:
     """
     Load the niche cluster assignment for each niche level.
     Args:
         options: Values, the options from optparse
     Returns:
-        pd.DataFrame: the niche cluster assignment for each niche level
+        Optional[pd.DataFrame]: the niche cluster assignment for each niche level
     """
 
     niche_level_niche_cluster_assign_file = f'{options.GNN_dir}/niche_level_niche_cluster.csv.gz'
     if not os.path.isfile(niche_level_niche_cluster_assign_file):
         niche_level_niche_cluster_assign_file = f'{options.GNN_dir}/niche_level_niche_cluster.csv'
     if not os.path.isfile(niche_level_niche_cluster_assign_file):  # skip if file not exist
-        raise FileNotFoundError(
-            f"Cannot find niche level niche cluster assign file: {niche_level_niche_cluster_assign_file}.")
+        warning(f"Cannot find niche level niche cluster assign file: {niche_level_niche_cluster_assign_file}.")
+        return None
 
     return pd.read_csv(niche_level_niche_cluster_assign_file, index_col=0)
 
 
-def load_cell_level_niche_cluster_assign(options: Values) -> pd.DataFrame:
+def load_cell_level_niche_cluster_assign(options: Values) -> Optional[pd.DataFrame]:
     """
     Load the niche cluster assignment for each cell level.
     Args:
         options: Values, the options from optparse
     Returns:
-        pd.DataFrame: the niche cluster assignment for each cell level
+        Optional[pd.DataFrame]: the niche cluster assignment for each cell level
     """
 
     cell_level_niche_cluster_assign_file = f'{options.GNN_dir}/cell_level_niche_cluster.csv.gz'
     if not os.path.isfile(cell_level_niche_cluster_assign_file):
         cell_level_niche_cluster_assign_file = f'{options.GNN_dir}/cell_level_niche_cluster.csv'
     if not os.path.isfile(cell_level_niche_cluster_assign_file):  # skip if file not exist
-        raise FileNotFoundError(
-            f"Cannot find cell level niche cluster assign file: {cell_level_niche_cluster_assign_file}.")
+        warning(f"Cannot find cell level niche cluster assign file: {cell_level_niche_cluster_assign_file}.")
+        return None
 
     return pd.read_csv(cell_level_niche_cluster_assign_file, index_col=0)
 
 
-def load_niche_level_max_niche_cluster(options: Values) -> pd.DataFrame:
+def load_niche_level_max_niche_cluster(options: Values) -> Optional[pd.DataFrame]:
     """
     Load the max niche cluster assignment for each niche level.
     Args:
         options: Values, the options from optparse
     Returns:
-        pd.DataFrame: the max niche cluster assignment for each niche level
+        Optional[pd.DataFrame]: the max niche cluster assignment for each niche level
     """
 
     niche_level_max_niche_cluster_file = f'{options.GNN_dir}/niche_level_max_niche_cluster.csv.gz'
     if not os.path.isfile(niche_level_max_niche_cluster_file):
         niche_level_max_niche_cluster_file = f'{options.GNN_dir}/niche_level_max_niche_cluster.csv'
     if not os.path.isfile(niche_level_max_niche_cluster_file):  # skip if file not exist
-        raise FileNotFoundError(
-            f"Cannot find niche level max niche cluster file: {niche_level_max_niche_cluster_file}.")
+        warning(f"Cannot find niche level max niche cluster file: {niche_level_max_niche_cluster_file}.")
+        return None
 
     return pd.read_csv(niche_level_max_niche_cluster_file, index_col=0)
 
 
-def load_cell_level_max_niche_cluster(options: Values) -> pd.DataFrame:
+def load_cell_level_max_niche_cluster(options: Values) -> Optional[pd.DataFrame]:
     """
     Load the max niche cluster assignment for each cell level.
     Args:
         options: Values, the options from optparse
     Returns:
-        pd.DataFrame: the max niche cluster assignment for each cell level
+        Optional[pd.DataFrame]: the max niche cluster assignment for each cell level
     """
 
     cell_level_max_niche_cluster_file = f'{options.GNN_dir}/cell_level_max_niche_cluster.csv.gz'
     if not os.path.isfile(cell_level_max_niche_cluster_file):
         cell_level_max_niche_cluster_file = f'{options.GNN_dir}/cell_level_max_niche_cluster.csv'
     if not os.path.isfile(cell_level_max_niche_cluster_file):  # skip if file not exist
-        raise FileNotFoundError(f"Cannot find cell level max niche cluster file: {cell_level_max_niche_cluster_file}.")
+        warning(f"Cannot find cell level max niche cluster file: {cell_level_max_niche_cluster_file}.")
+        return None
 
     return pd.read_csv(cell_level_max_niche_cluster_file, index_col=0)
 
@@ -224,13 +227,13 @@ class AnaData:
 
     @property
     def train_loss(self):
-        if not hasattr(self, '_train_loss'):
+        if not hasattr(self, '_train_loss') or self._train_loss is None:
             self._train_loss = load_loss_record_data(self.options)
         return self._train_loss
 
     @property
     def cell_type_codes(self) -> pd.DataFrame:
-        if not hasattr(self, '_cell_type_codes'):
+        if not hasattr(self, '_cell_type_codes') or self._cell_type_codes is None:  # type: ignore
             self._cell_type_codes = pd.read_csv(f'{self.options.NN_dir}/cell_type_code.csv', index_col=0)
         return self._cell_type_codes
 
@@ -247,60 +250,74 @@ class AnaData:
             data_df = pd.concat([data_df, cell_type_composition_df])
         self._cell_type_composition = data_df.loc[self.meta_data_df.index]
 
-    def _load_NT_score(self) -> None:
-        NTScore_df = pd.read_csv(f'{self.options.NT_dir}/NTScore.csv.gz', index_col=0)
-        self._NT_score = NTScore_df.loc[self.meta_data_df.index]
-
     @property
     def cell_type_composition(self) -> DataFrame:
-        if not hasattr(self, '_cell_type_composition'):
+        if not hasattr(self, '_cell_type_composition') or self._cell_type_composition is None:
             self._load_cell_type_composition()
         return self._cell_type_composition
 
+    def _load_NT_score(self) -> Optional[DataFrame]:
+        if not os.path.isfile(f'{self.options.NT_dir}/NTScore.csv.gz'):
+            warning(f"Cannot find NT score file: {self.options.NT_dir}/NTScore.csv.gz.")
+            return None
+        NTScore_df = pd.read_csv(f'{self.options.NT_dir}/NTScore.csv.gz', index_col=0)
+        return NTScore_df.loc[self.meta_data_df.index]
+
     @property
-    def NT_score(self) -> DataFrame:
-        if not hasattr(self, '_NT_score'):
-            self._load_NT_score()
+    def NT_score(self) -> Optional[DataFrame]:
+        if not hasattr(self, '_NT_score') or self._NT_score is None: # type: ignore
+            self._NT_score = self._load_NT_score()
         return self._NT_score
 
     @property
-    def niche_cluster_connectivity(self) -> np.ndarray:
-        if not hasattr(self, '_niche_cluster_connectivity'):
-            # FileNotFoundError will be raised if the file does not exist
+    def niche_cluster_connectivity(self) -> Optional[np.ndarray]:
+        if not hasattr(self, '_niche_cluster_connectivity') or self._niche_cluster_connectivity is None:  # type: ignore
             self._niche_cluster_connectivity = load_niche_cluster_connectivity(self.options)
         return self._niche_cluster_connectivity
 
     @property
-    def niche_cluster_score(self) -> np.ndarray:
-        if not hasattr(self, '_niche_cluster_score'):
-            # FileNotFoundError will be raised if the file does not exist
+    def niche_cluster_score(self) -> Optional[np.ndarray]:
+        if not hasattr(self, '_niche_cluster_score') or self._niche_cluster_score is None:  # type: ignore
             self._niche_cluster_score = load_niche_cluster_score(self.options)
         return self._niche_cluster_score
 
     @property
-    def niche_level_niche_cluster_assign(self) -> pd.DataFrame:
-        if not hasattr(self, '_niche_level_niche_cluster_assign'):
+    def niche_level_niche_cluster_assign(self) -> Optional[pd.DataFrame]:
+        if not hasattr(
+                self,
+                '_niche_level_niche_cluster_assign') or self._niche_level_niche_cluster_assign is None:  # type: ignore
             self._niche_level_niche_cluster_assign = load_niche_level_niche_cluster_assign(self.options)
+            if self._niche_level_niche_cluster_assign is None:
+                return None
             try:
-                self._niche_level_niche_cluster_assign = self._niche_level_niche_cluster_assign.loc[self.meta_data_df.index]
+                self._niche_level_niche_cluster_assign = self._niche_level_niche_cluster_assign.loc[
+                    self.meta_data_df.index]
             except:
                 pass
         return self._niche_level_niche_cluster_assign
 
     @property
-    def cell_level_niche_cluster_assign(self) -> pd.DataFrame:
-        if not hasattr(self, '_cell_level_niche_cluster_assign'):
+    def cell_level_niche_cluster_assign(self) -> Optional[pd.DataFrame]:
+        if not hasattr(
+                self,
+                '_cell_level_niche_cluster_assign') or self._cell_level_niche_cluster_assign is None:  # type: ignore
             self._cell_level_niche_cluster_assign = load_cell_level_niche_cluster_assign(self.options)
+            if self._cell_level_niche_cluster_assign is None:
+                return None
             try:
-                self._cell_level_niche_cluster_assign = self._cell_level_niche_cluster_assign.loc[self.meta_data_df.index]
+                self._cell_level_niche_cluster_assign = self._cell_level_niche_cluster_assign.loc[
+                    self.meta_data_df.index]
             except:
                 pass
         return self._cell_level_niche_cluster_assign
 
     @property
-    def niche_level_max_niche_cluster(self) -> pd.DataFrame:
-        if not hasattr(self, '_niche_level_max_niche_cluster'):
+    def niche_level_max_niche_cluster(self) -> Optional[pd.DataFrame]:
+        if not hasattr(self,
+                       '_niche_level_max_niche_cluster') or self._niche_level_max_niche_cluster is None:  # type: ignore
             self._niche_level_max_niche_cluster = load_niche_level_max_niche_cluster(self.options)
+            if self._niche_level_max_niche_cluster is None:
+                return None
             try:
                 self._niche_level_max_niche_cluster = self._niche_level_max_niche_cluster.loc[self.meta_data_df.index]
             except:
@@ -308,9 +325,12 @@ class AnaData:
         return self._niche_level_max_niche_cluster
 
     @property
-    def cell_level_max_niche_cluster(self) -> pd.DataFrame:
-        if not hasattr(self, '_cell_level_max_niche_cluster'):
+    def cell_level_max_niche_cluster(self) -> Optional[pd.DataFrame]:
+        if not hasattr(self,
+                       '_cell_level_max_niche_cluster') or self._cell_level_max_niche_cluster is None:  # type: ignore
             self._cell_level_max_niche_cluster = load_cell_level_max_niche_cluster(self.options)
+            if self._cell_level_max_niche_cluster is None:
+                return None
             try:
                 self._cell_level_max_niche_cluster = self._cell_level_max_niche_cluster.loc[self.meta_data_df.index]
             except: